requirements.txt
ipywidgets
dvc
dvc-gdrive
hyperopt
mlflow
catboost
python-dotenv
fastapi
pandas
prefect
tqdm
requests
joblib
pyarrow
psycopg
psycopg-binary
evidently
numpy
scikit-learn
jupyter
matplotlib
xgboost
ydata-profiling